requirements.txt
wheel
numpy
pandas
shapely
patool
fastapi
uvicorn[standard]
python-dotenv
pydantic
python-multipart
psycopg[binary]
sentinelsat
matplotlib
six
rtree
pipwin
rasterio
pyproj
geopandas
fiona









